Move DataSourceWrapper from pipeline-core module to pipeline-spi module
